chore: move config to separate file config.go (#16832)
